Mailer Work, send link with public url

+++ app/controllers/maps_controller.rb
@@ -1,19 +1,19 @@
 class MapsController < ApplicationController
+  skip_after_action :verify_authorized, only: :create
+
   def new
     @map = Map.new
   end
 
-  def show
-    @map = Map.find(params[:id])
-    @creator = @map.user
-  end
 
   def show_public
     @map = Map.find_by(public_id: params[:public_id])
+
   end
 
   def edit
     @map = Map.find(params[:id])
+    authorize @map
     @creator = @map.user
   end
 
@@ -28,14 +28,16 @@ class MapsController < ApplicationController
   def update
     if params[:map_name].present?
       new_name = params[:map_name]
-      map =  Map.find(params[:id])
-      map.update_attributes(name: new_name )
+      @map =  Map.find(params[:id])
+      authorize @map
+      @map.update_attributes(name: new_name )
       redirect_to edit_map_path
     end
   end
 
   def duplicate
     @map_origin = Map.find(params[:id])
+    authorize @map_origin
     @map_duplicate_name = @map_origin.name + "-copy-" + Time.now.to_formatted_s(:number)
     current_user_id =  current_user.id
     @map_duplicate = Map.new(name: @map_duplicate_name,
@@ -45,16 +47,25 @@ class MapsController < ApplicationController
                             latitude: @map_origin.latitude
                             )
     @map_duplicate.save
-    @map_duplicate.duplicate(@map_origin)
+    @map_duplicate.duplicate_map_places(@map_origin)
     redirect_to edit_map_path(@map_duplicate)
   end
 
   def destroy
     map = Map.find(params[:id])
+    authorize map
     map.destroy
     redirect_to profil_user_path(current_user)
   end
 
+
+  def map_share
+    info_email = params
+    map = Map.find(params[:id])
+    authorize map
+    SharingMapMailer.share_a_map(info_email).deliver_now
+    
+  end
 
 
 end
